Resolved dependency on ffmpeg-python
When building in rawhide, the package is not provided, and may not
be available for f42.  However, rebuilding the f41 version works
fine so a copy has been place in a COPR until package status gets
resolved, either by someone adopting it or by gramps-web-api using
some other module instead.

Signed-off-by: Al Stone <ahs3@ahs3.net>