Attempted %check in python-limits 4.0.1-2
Deferring this for now until we can get upstream to straighten
out the test cases.  This module uses python3-etcd3 in the test
cases but that module cannot be tested, nor can it actually be
used in the limits module since it relies on a deprecated protobuf
module.  So, we can build the test cases, but they will not run.

Signed-off-by: Al Stone <ahs3@ahs3.net>