Mostly working python3-gramps-webapi 2.7.0-2
Packaging seems to be working, and hand-testing indicates that the
webapi will at least start (NB: the run_tests.sh script automates
this basic test).

Attempted to get %check working, but there are a huge number of
dependencies for that, some of them very large (e.g., python3-openai
and python3-torch).  Will need to investigate this over time in order
to automate it.

For now, the basics appear to be working.

Signed-off-by: Al Stone <ahs3@ahs3.net>